feat(client, server, common): base structure for chat and rooms

+++ server/src/main/java/org/cucumber/server/repositories/IBasicRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public interface IBasicRepository<T> {
 
-    T findById(Long id);
+    T findById(Integer id);
 
     List<T> findAll();
